Fixed bugs
Fixed the bug that causes stackoverflow when raising to the power of the minimum integer, fixed the bug where Complex.equals(null) results in a NullPointerException (which was caused by having an equals(Complex) and an equals(Object) function, causing null to be cast to a complex and then a NullPointerException). Made Complex implement Cloneable, just because it's a good idea. And, lastly, changed the small detail in the README files where I say "modular exponentiation" when I really mean "integer exponentiation", because I guess when I first made this library I wasn't yet aware of what "modular" meant.